Check that x is a valid path

Check that x is a valid path, possibly containing a valid filename.

Usage

is_path(path)

Arguments

path

character string with the path, possibly containing a valid filename.

Value

TRUE or FALSE indicating if path is a valid path, possibly containing a valid filename.

Details

is_path() is intended to be used to check for valid paths before creating a directory or a file. Therefore it imposes the following restrictions:

path should not contain the characters ", *, ?, |, <, >, nor any of the control characters (ASCII octal codes 000 through 037 and 177, see help("regex")).

path components (i.e., parts separated by file separators / or \\) should not be the Windows-reserved terms CON, PRN, AUX, NUL, COM<non-zero digit>, LPT<non-zero digit>, case-insensitive variants of these names, or these names followed by an extension.

path components should not end with a space.

path components should not end with a dot, with the exception of "." and ".." that are allowed as first component to indicate the working directory and the parent directory, respectively.

path should not point to tempdir(): a temporary subdirectory should be used instead (see progutils::create_tempdir()).

If path contains a file extension (or compression extension, the current implementation does not distinguish those from each other), the part after the last slash is considered the filename, which should adhere to the restrictions listed above, and in addition should not contain : nor start with a space or a hyphen (-), while it might contain the Windows-reserved terms given in the second point above.

These restrictions path consider characters and words that are not allowed in Windows and thus would lead to an error when used to create a directory or file; and characters that are silently removed in Windows and thus would lead to a mismatch between the created directory and the returned path when used to create a directory.

is_path() allows some patterns that will not occur in real (i.e., existing) paths or filenames:

path does not have to contain a file separator (i.e., / or \\). This makes it possible to use is_path() to check that input to fs::path() only contains allowed characters.

path does not have to point to an existing directory (see the previous point).

path might contain repeated file separators (e.g., // or \\\\): these will be treated as if they were only a single file separator.

path might contain trailing file separators, even though these might be ignored or removed in some operations (e.g., they are removed by file.path() and fs::path()) .

Programming notes

The file separator is a backslash (\) on Windows but a forward slash (/) on other operating systems (.Platform$file.sep gives the file separator used on the current platform). Furthermore, the backslash is used as escape character in R, such that backslashes need to be escaped in R code. Thus, a check on the presence of repeated slashes and backslashes in string string would use grepl(pattern = "//", x = string, fixed = TRUE) and grepl(pattern = "\\\\", x = string, fixed = TRUE). The message to point out their presence would be written as message("Repeated '/' or '\\'") which would be printed as Repeated '/' or '\'. This makes it cumbersome to get the correct type and number of slashes to compare with the path recorded in a message, such that it is more robust to check only for fixed parts of the message (e.g., "Repeated"), possibly followed by a check like tinytest::expect_true(dir.exists(string)).

On MacOS, the output of tempdir() is preceded by duplicated forward slashes (e.g., /var/[...]/T//RtmpxC2Fyl/working_dir/RtmpdnqgUR) which led to spurious warnings in earlier versions of is_path() that warned about duplicated file separators.

References

Naming files, paths, and namespaces from Microsoft

Entries Filename, Portable filenames and Pathname from the Posix standard POSIX.1-2024

Comparison of file systems from Wikipedia

See also

Examples

is_path(getwd())
#> [1] TRUE
is_path(fs::path_wd("abcd"))
#> [1] TRUE
is_path(fs::path_wd("ab|cd"))
#> Warning: 'path' ('fs::path_wd("ab|cd")') should not contain '"', '*', '?', '|', '<' or '>':
#> /home/runner/work/checkinput/checkinput/docs/reference/ab|cd
#> [1] FALSE

is_path(fs::path_wd("abcd.txt"))
#> [1] TRUE
is_path(fs::path_wd("abcd.txt.gz"))
#> [1] TRUE
is_path(fs::path_wd("abcd.gz"))
#> [1] TRUE

is_path(fs::path_wd("ab:cd.txt"))
#> Warning: 'filename' should not contain ':':
#> ab:cd.txt
#> [1] FALSE
is_path(fs::path_wd("ab|cd.txt"))
#> Warning: 'path' ('fs::path_wd("ab|cd.txt")') should not contain '"', '*', '?', '|', '<' or '>':
#> /home/runner/work/checkinput/checkinput/docs/reference/ab|cd.txt
#> [1] FALSE